Avoid stylesheets in connection and login widgets
Set relative font size in code instead of specifying a 'larger' font
size using stylesheets.
