+++ src/desktop-remote/connectionwidget.ui
@@ -17,7 +17,7 @@
    <item row="0" column="1">
     <spacer name="verticalSpacer">
      <property name="orientation">
-      <enum>Qt::Vertical</enum>
+      <enum>Qt::Orientation::Vertical</enum>
      </property>
      <property name="sizeHint" stdset="0">
       <size>
@@ -30,7 +30,7 @@
    <item row="1" column="0">
     <spacer name="horizontalSpacer">
      <property name="orientation">
-      <enum>Qt::Horizontal</enum>
+      <enum>Qt::Orientation::Horizontal</enum>
      </property>
      <property name="sizeHint" stdset="0">
       <size>
@@ -44,9 +44,6 @@
     <layout class="QVBoxLayout" name="verticalLayout">
      <item>
       <widget class="QLabel" name="titleLabel">
-       <property name="styleSheet">
-        <string notr="true">font:10pt;</string>
-       </property>
        <property name="text">
         <string>Connect to your PMP server</string>
        </property>
@@ -76,7 +73,7 @@
        <item>
         <spacer name="horizontalSpacer_4">
          <property name="orientation">
-          <enum>Qt::Horizontal</enum>
+          <enum>Qt::Orientation::Horizontal</enum>
          </property>
          <property name="sizeHint" stdset="0">
           <size>
@@ -98,7 +95,7 @@
      <item>
       <spacer name="horizontalSpacer_3">
        <property name="orientation">
-        <enum>Qt::Horizontal</enum>
+        <enum>Qt::Orientation::Horizontal</enum>
        </property>
        <property name="sizeHint" stdset="0">
         <size>
@@ -117,7 +114,7 @@
         <item row="0" column="1">
          <widget class="QLineEdit" name="serverLineEdit"/>
         </item>
-        <item row="2" column="1" alignment="Qt::AlignRight">
+        <item row="2" column="1" alignment="Qt::AlignmentFlag::AlignRight">
          <widget class="QPushButton" name="connectButton">
           <property name="text">
            <string>Connect</string>
@@ -149,7 +146,7 @@
    <item row="1" column="2">
     <spacer name="horizontalSpacer_2">
      <property name="orientation">
-      <enum>Qt::Horizontal</enum>
+      <enum>Qt::Orientation::Horizontal</enum>
      </property>
      <property name="sizeHint" stdset="0">
       <size>
@@ -162,7 +159,7 @@
    <item row="2" column="1">
     <spacer name="verticalSpacer_2">
      <property name="orientation">
-      <enum>Qt::Vertical</enum>
+      <enum>Qt::Orientation::Vertical</enum>
      </property>
      <property name="sizeHint" stdset="0">
       <size>
