+++ src/desktop-remote/mainwidget.cpp
@@ -40,6 +40,7 @@
 #include "scoreformatdelegate.h"
 #include "trackinfodialog.h"
 #include "userforstatisticsdisplay.h"
+#include "widgetutils.h"
 
 #include <cmath>
 
@@ -53,28 +54,6 @@ using namespace PMP::UnicodeChars;
 
 namespace PMP
 {
-    namespace
-    {
-        void setRelativeFontSize(QLabel* label, double scaleFactor)
-        {
-            QFont font = label->font();
-            double currentPointSize = font.pointSize();
-
-            /* If the font size is not explicitly set (pointSize <= 0), calculate it
-               from QFontMetrics */
-            if (currentPointSize <= 0)
-            {
-                QFontMetrics metrics(font);
-
-                /* approximation to convert height to point size */
-                currentPointSize = metrics.height() / 1.2;
-            }
-
-            font.setPointSizeF(currentPointSize * scaleFactor);
-            label->setFont(font);
-        }
-    }
-
     MainWidget::MainWidget(QWidget* parent)
      : QWidget(parent),
         _ui(new Ui::MainWidget),
@@ -93,7 +72,7 @@ namespace PMP
         _ui->splitter->setStretchFactor(0, 4);
         _ui->splitter->setStretchFactor(1, 8);
 
-        setRelativeFontSize(_ui->artistTitleLabel, 1.3);
+        WidgetUtils::setRelativeFontSize(_ui->artistTitleLabel, 1.3);
 
         auto trackTimeLabel = ClickableLabel::replace(_ui->positionLabel);
         auto trackTimeValueLabel = ClickableLabel::replace(_ui->positionValueLabel);
